Added unsigned long int conversion to handler size because builds in Travis vs. locally represent the handler size as either unsigned int or unsigned long int

+++ main.cc
@@ -32,7 +32,7 @@ int main(int argc, char* argv[]) {
         // Parse the echo and static file request handlers from the config file
         std::vector<std::unique_ptr<http::handler> > handlers;
         server_parser.ParseRequestHandlers(handlers);
-        printf("%d handlers parsed\n", handlers.size());
+        printf("%lu handlers parsed\n", (unsigned long int)(handlers.size()));
 
         // Start the server
         boost::asio::io_service io_service;
